+++ src/zopflipng/zopflipng_lib.h
@@ -29,6 +29,7 @@ strategies.
 
 #ifdef __cplusplus
 
+#include "../zopfli/defines.h"
 #include <string>
 #include <vector>
 #include <algorithm>
@@ -281,7 +282,7 @@ struct ZopfliPNGOptions {
   0x0080 - Use expensive fixed block calculations in splitter,
   0x0100 - Use Restore Points.
   */
-  int mode;
+  unsigned long mode;
 
   /*
   Iterate multiple dynamic blocks at once using pthreads, aka.
